+++ src/net/kopeph/ld31/menu/MenuWidget.java
@@ -15,7 +15,7 @@ public abstract class MenuWidget {
 
 	protected final PApplet context = LD31.getContext();
 
-	protected int xPos, yPos, width, height;
+	protected double xPos, yPos, width, height;
 	protected int xAnchor = ANCHOR_LEFT, yAnchor = ANCHOR_TOP;
 
 	private int curCtxWidth, curCtxHeight;
@@ -51,18 +51,18 @@ public abstract class MenuWidget {
 		curCtxHeight = context.height;
 
 		switch(xAnchor) {
-		case ANCHOR_CENTER: xPos += dCtxWidth/2; break;
-		case ANCHOR_LEFT: 						 break;
-		case ANCHOR_RIGHT:  xPos += dCtxWidth; 	 break;
-	    case ANCHOR_FILL:  width += dCtxWidth; 	 break;
+		case ANCHOR_CENTER: xPos += dCtxWidth/2.0; break;
+		case ANCHOR_LEFT: 						   break;
+		case ANCHOR_RIGHT:  xPos += dCtxWidth; 	   break;
+	    case ANCHOR_FILL:  width += dCtxWidth; 	   break;
 		}
 
 
 		switch(yAnchor) {
-		case ANCHOR_CENTER: yPos += dCtxHeight/2; break;
-		case ANCHOR_TOP: 						  break;
-		case ANCHOR_BOTTOM: yPos += dCtxHeight;   break;
-	    case ANCHOR_FILL: height += dCtxHeight;   break;
+		case ANCHOR_CENTER: yPos += dCtxHeight/2.0; break;
+		case ANCHOR_TOP: 						    break;
+		case ANCHOR_BOTTOM: yPos += dCtxHeight;     break;
+	    case ANCHOR_FILL: height += dCtxHeight;     break;
 		}
 	}
